// src/app/interfaces/user.ts
export interface User {
  allowedToAccess?: boolean;
  avatar?: string;
  capability?: number;
  createdAt?: string;
  email?: string;
  emailValidated?: boolean;
  firstName?: string;
  fullName?: string;
  id?: string;
  isMale?: boolean;
  jwtValidatedAt?: string;
  lastName?: string;
  online?: boolean;
  updatedAt?: string;
  username?: string;
  _id?: string;
}